requirements.txt
yfinance==0.2.37
python-telegram-bot==20.8
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
